sectioned off cart into components
